// clients/etx/src/core/transaction/types.ts
export const TRANSACTION_TYPE = {
  Legacy: 'legacy',
  Eip2930: 'eip2930',
  Eip1559: 'eip1559',
} as const

export type TransactionType = (typeof TRANSACTION_TYPE)[keyof typeof TRANSACTION_TYPE]
// ...
export interface ISignableTransaction {
  readonly type: TransactionType
  readonly chainId: ChainId
  readonly from: Address
  readonly to: Address | null
  readonly value: Wei
  readonly data: HexString
  readonly nonce: number
  readonly gasLimit: bigint

  /** Заполняется для EIP-1559. */
  readonly maxFeePerGas: bigint | null
  readonly maxPriorityFeePerGas: bigint | null

  /** Заполняется для транзакций прежнего формата. */
  readonly gasPrice: bigint | null
}
// ...
export interface ITypedData {
  readonly domain: ITypedDataDomain
  readonly types: Readonly<Record<string, readonly ITypedDataField[]>>
  readonly primaryType: string
  readonly message: Readonly<Record<string, unknown>>
}

export interface ITypedDataDomain {
  readonly name?: string
  readonly version?: string
  readonly chainId?: ChainId
  readonly verifyingContract?: Address
  readonly salt?: HexString
}

export interface ITypedDataField {
  readonly name: string
  readonly type: string
}